+++ tests/Provider/FakeFileProvider.php
@@ -72,6 +72,6 @@ class FakeFileProvider extends \Psalm\Provider\FileProvider
     public function registerFile($file_path, $file_contents)
     {
         $this->fake_files[$file_path] = $file_contents;
-        $this->fake_file_times[$file_path] = (float) microtime(true);
+        $this->fake_file_times[$file_path] = microtime(true);
     }
 }
